requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flake8==5.0.4
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
identify==2.5.18
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pytest==7.2.1
PyYAML==6.0
requests==2.28.2
SQLAlchemy==2.0.4
stripe==5.2.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.19.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.14.0
